+++ packages/espressocash_app/lib/features/activities/services/pending_activities_repository.dart
@@ -8,8 +8,8 @@ import '../../../data/db/db.dart';
 import '../../outgoing_direct_payments/data/repository.dart';
 import '../../outgoing_link_payments/data/repository.dart';
 import '../../payment_request/data/repository.dart';
-import '../../ramp/data/on_ramp_order_service.dart';
 import '../../ramp/services/off_ramp_order_service.dart';
+import '../../ramp/services/on_ramp_order_service.dart';
 import '../../tokens/token_list.dart';
 import '../../transaction_request/service/tr_service.dart';
 import '../data/activity_builder.dart';
