+++ packages/espressocash_app/lib/features/ramp/widgets/on_ramp_order_details.dart
@@ -2,13 +2,13 @@ import 'package:flutter/widgets.dart';
 
 import '../../../di.dart';
 import '../../ramp_partner/models/ramp_partner.dart';
-import '../data/on_ramp_order_service.dart';
 import '../models/ramp_watcher.dart';
 import '../partners/kado/services/kado_on_ramp_order_watcher.dart';
 import '../partners/moneygram/service/onramp/moneygram_on_ramp_order_watcher.dart';
 import '../partners/scalex/services/scalex_on_ramp_order_watcher.dart';
+import '../services/on_ramp_order_service.dart';
 
-export '../data/on_ramp_order_service.dart' show OnRampOrder;
+export '../services/on_ramp_order_service.dart' show OnRampOrder;
 
 typedef OnRampOrderDetailsBuilder = Widget Function(
   BuildContext context,
